+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#16](https://github.com/LaxarJS/ax-input-control/issues/16): ignore `title` attribute when showing validation message
+
 
 ## v2.0.0-alpha.5